chore(hooks): organizing use hooks files remove duplication
